[Testing]

+++ secure/net/src/lib.rs
@@ -501,12 +501,12 @@ impl NetworkStream {
     /// This wraps around that buffer and blocks until all the data has been pushed.
     fn write_all(&mut self, data: &[u8]) -> Result<(), Error> {
         let mut unwritten = data;
-        let mut total_written: u8 = 0;
+        let mut total_written: u64 = 0;
 
         while !unwritten.is_empty() {
             let written = self.stream.write(unwritten)?;
             total_written = total_written
-                .checked_add(written as u8)
+                .checked_add(written as u64)
                 .ok_or_else(|| Error::OverflowError("write_all::total_written".into()))?;
             unwritten = &data[total_written as usize..];
         }
